Revert "Song page를 SongDetail로 변경"
This reverts commit 3a7e7c376f03611489d8d191d0185de9ca7a8b63.
